webapp/requirements.txt
astroid==1.6.1
boto3==1.6.3
botocore==1.9.3
Django==1.11
django-filter==1.1.0
django-localflavor==2.0
django-money==0.12.3
django-seed==0.1.8
django-storages==1.6.5
django-widget-tweaks==1.4
docutils==0.14
Faker==0.8.11
isort==4.3.3
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
pkg-resources==0.0.0
py-moneyed==0.7.0
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.13
six==1.11.0
text-unidecode==1.2
wrapt==1.10.11
